Break update detection because forks are not supposed to be updated with upstream SMH.
its not like I don't want this to work well, I really do. But I don't understand JS very well, and the fact that this works at all is amazing to me. I'm yeeting the update URL so that BD doesn't constantly nag me install the latest, when its already the latest version of my fork.